TermsOfService.js -> TermsOfService.ts
